Further clean up
